+++ griptape/drivers/prompt/dummy_prompt_driver.py
@@ -6,7 +6,7 @@ from attrs import Factory, define, field
 
 from griptape.common import observable
 from griptape.drivers import BasePromptDriver
-from griptape.exceptions import DummyException
+from griptape.exceptions import DummyError
 from griptape.tokenizers import DummyTokenizer
 
 if TYPE_CHECKING:
@@ -22,8 +22,8 @@ class DummyPromptDriver(BasePromptDriver):
 
     @observable
     def try_run(self, prompt_stack: PromptStack) -> Message:
-        raise DummyException(__class__.__name__, "try_run")
+        raise DummyError(__class__.__name__, "try_run")
 
     @observable
     def try_stream(self, prompt_stack: PromptStack) -> Iterator[DeltaMessage]:
-        raise DummyException(__class__.__name__, "try_stream")
+        raise DummyError(__class__.__name__, "try_stream")