+++ griptape/tasks/prompt_task.py
@@ -103,6 +103,8 @@ class PromptTask(RuleMixin, BaseTask):
             return task_input
         elif isinstance(task_input, Callable):
             return self._process_task_input(task_input(self))
+        elif isinstance(task_input, ListArtifact):
+            return ListArtifact([self._process_task_input(elem) for elem in task_input.value])
         elif isinstance(task_input, BaseArtifact):
             return task_input
         elif isinstance(task_input, (list, tuple)):