+++ tests/unit/tasks/test_variation_image_generation_task.py
@@ -27,20 +27,20 @@ class TestVariationImageGenerationTask:
         assert task.input.value == list(input_tuple)
 
     def test_callable_input(self, text_artifact: TextArtifact, image_artifact: ImageArtifact):
-        input = [text_artifact, image_artifact]
+        artifacts = [text_artifact, image_artifact]
 
-        def callable(task: BaseTask) -> ListArtifact:
-            return ListArtifact(input)
+        def callable_input(task: BaseTask) -> ListArtifact:
+            return ListArtifact(artifacts)
 
-        task = VariationImageGenerationTask(callable, image_generation_engine=Mock())
+        task = VariationImageGenerationTask(callable_input, image_generation_engine=Mock())
 
-        assert task.input.value == input
+        assert task.input.value == artifacts
 
     def test_list_input(self, text_artifact: TextArtifact, image_artifact: ImageArtifact):
-        input = [text_artifact, image_artifact]
-        task = VariationImageGenerationTask(ListArtifact(input), image_generation_engine=Mock())
+        artifact_input = [text_artifact, image_artifact]
+        task = VariationImageGenerationTask(ListArtifact(artifact_input), image_generation_engine=Mock())
 
-        assert task.input.value == input
+        assert task.input.value == artifact_input
 
     def test_bad_input(self, image_artifact):
         with pytest.raises(ValueError):